change: moved editor functions to useEditor hook
