+++ client/src/stores/editorState.ts
@@ -10,6 +10,8 @@ interface EditorState {
   } | null;
   collidingId: string | null;
   originPosition: Position | null;
+  currentEditorId: string | null;
+  setCurrentEditorId: (id: string | null) => void;
   setItems: (items: EditorItem[]) => void;
   addNewItem: (item: EditorItem) => void;
   deleteItem: (itemId: string) => void;
@@ -50,6 +52,8 @@ export const useEditorStore = create<EditorState>((set) => ({
   collidingId: null,
   selectedItem: null,
   originPosition: null,
+  currentEditorId: localStorage.getItem("recentEditorId"),
+  setCurrentEditorId: (value) => set({ currentEditorId: value }),
   setItems: (items) => set({ items }),
   addNewItem: (item) => set((state) => ({ items: [...state.items, item] })),
   deleteItem: (itemId) =>
